pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.5.4",
    "langchain>=0.3.21",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "wikipedia>=1.4.0",
]
